+++ src/V1/Tests/TestFiles/StartupSqlServer.cs
@@ -18,6 +18,7 @@ namespace ServiceBricks.Xunit
             services.AddSingleton(Configuration);
             services.AddServiceBricks(Configuration);
             services.AddServiceBricksLoggingSqlServer(Configuration);
+            services.AddServiceBricksComplete(Configuration);
 
             // Remove all background tasks/timers for unit testing
             var logtimer = services.Where(x => x.ImplementationType == typeof(CustomLoggerWriteMessageTimer)).FirstOrDefault();
@@ -27,8 +28,6 @@ namespace ServiceBricks.Xunit
             // Register TestManager
             services.AddScoped<ITestManager<LogMessageDto>, LogMessageTestManager>();
             services.AddScoped<ITestManager<WebRequestMessageDto>, WebRequestMessageTestManager>();
-
-            services.AddServiceBricksComplete(Configuration);
         }
 
         public virtual void Configure(IApplicationBuilder app)